add delete method for board backend and connect it with workspace backend

+++ back-end/src/main/java/group16/backend/workspace/repository/WorkspaceRepo.java
@@ -1,10 +1,12 @@
 package group16.backend.workspace.repository;
 
 import group16.backend.workspace.entity.Workspace;
+
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkspaceRepo extends JpaRepository<Workspace, Integer> {
+public interface WorkspaceRepo extends JpaRepository<Workspace, Long> {
 
 }
